feat(interpreter): add dynamic sublinear
GitOrigin-RevId: 7de54fe7433fcad91d36b91297c87ccf9eaa5fe5

+++ imperative/python/src/tensor.cpp
@@ -901,7 +901,7 @@ void init_tensor(py::module m) {
     }
 
     m.def("set_option",
-          [](std::string name, int value){ interpreter_for_py->set_option(name, value); });
+          [](std::string name, size_t value){ interpreter_for_py->set_option(name, value); });
     m.def("get_option",
           [](std::string name){ return interpreter_for_py->get_option(name); });
     m.def("_set_swap_flag",
